[FP-49] Refactor: 페이지네이션 수정 및 체험 상세 페이지 리팩토링
